requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-vlc==3.0.21203
requests==2.32.3
shiboken6==6.9.1
urllib3==2.4.0
